내부 클래스 사용 (1) - 전략 클래스를 메소드 내부에 선언
* 외부에 선언된 변수를 사용할 수 있다 -> 생성자/멤버변수 필요 없음.
* 별도의 클래스 파일을 만들지 않아도 된다.
